Database auto upgrade

+++ app/src/main/java/ee/ttu/schedule/service/DatabaseHandler.java
@@ -16,7 +16,7 @@ import java.util.Set;
 
 public class DatabaseHandler extends SQLiteOpenHelper {
 
-    private static final int DATABASE_VERSION = 1;
+    private static final int DATABASE_VERSION = 3;
     private static final String DATABASE_NAME = "scheduleManager.db";
     private static final String TABLE_SCHEDULE = "schedule";
 
@@ -50,8 +50,10 @@ public class DatabaseHandler extends SQLiteOpenHelper {
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
         // Drop older table if existed
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_SCHEDULE);
-        onCreate(db);
+        if(oldVersion < 3) {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_SCHEDULE);
+            onCreate(db);
+        }
     }
 
     public void addContact(Subject subject) {
